JAX->NumPy conversion after OperatorTerm action batch fixed. __reversed__ fixed. IndexError as iter stopper fixed.

+++ solax/quantum_core/bit_level_primitives/ladder_mappings.py
@@ -32,7 +32,7 @@ def map_with_ladseq(det_code, bit_posits, daggers):
     See also help(map_with_ladder).
     """
     valid = 1
-    daggers = jnp.array(daggers)
+    daggers = jnp.array(daggers, dtype=jnp.uint8)
     
     lower = 1
     upper = len(daggers) + 1
